+++ Core/Render/Common/Textures/FilterType.cs
@@ -1,3 +1,5 @@
+using System.ComponentModel;
+
 namespace Helion.Render.Common.Textures;
 
 public enum FilterType
@@ -5,4 +7,8 @@ public enum FilterType
     Nearest,
     Bilinear,
     Trilinear,
+    [Description("Nearest to bilinear")]
+    NeareastBilinear,
+    [Description("Nearest to trilinear")]
+    NearestTrilinear
 }